Add command pattern.
