+++ Commands/Command.hpp
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <string>
+
+class Command{
+public:
+    Command() = default;
+    virtual ~Command() = default;
+
+    virtual void run() = 0;
+    virtual std::string getName() = 0;
+};
