+++ Menu.hpp
@@ -1,51 +1,13 @@
 #pragma once
+
+#include <array>
 #include <iostream>
+#include <memory>
 #include "Database.hpp"
+#include "Commands/Command.hpp"
 
 class Menu {
 public:
     Menu() {}
     void run();
-
-private:
-    struct Option {
-        using CallMethod = void (Menu::*)();
-        CallMethod calledMethod;
-        std::string optionDescription;
-    };
-
-    const std::vector<Option> options_{
-        {&Menu::printMenu, "Show Menu"},
-        {&Menu::printAllRecords, "Print All Records"},
-        {&Menu::loadRecords, "Load Record"},
-        {&Menu::addStudent, "Add Student"},
-        {&Menu::searchOption, "Search"},
-        {&Menu::searchByPesel, "Search By Pesel"},
-        {&Menu::sortBySurname, "Sort By Surname"},
-        {&Menu::sortByPesel, "Sort By Pesel"},
-        {&Menu::deleteByPesel, "Delete By Pesel"},
-        {&Menu::deleteByIndexNumber, "Delete By Index"},
-        {&Menu::validatePeselNumber, "Validate PESEL"},
-        {&Menu::saveRecords, "Save Records"},
-        {&Menu::endProgram, "End Program"}};
-
-    void showMenu() const;
-    void printMenu() { showMenu(); }
-    void printAllRecords();
-    void loadRecords();
-    void addStudent();
-    void sortBySurname();
-    void sortByPesel();
-    void searchOption();
-    void searchByPesel();
-    void deleteByPesel();
-    void deleteByIndexNumber();
-    void validatePeselNumber();
-    void saveRecords();
-    void endProgram();
-
-    void printStudent(const Student& student) const;
-
-    bool menuQuit = false;
-    Database db_;
 };
